Swap to single main loop instead of interrupts
This allows storing state inside the HAL drivers (since they no longer
need to be able to be "stolen" by the ISRs), removes all `steal()`
methods, and should simplify the rest of the application logic (dealing
with events coming to/from the USB).

We can use WFE and SEVONPEND to continue to sleep until an interrupt
occurs. Increased response latency should be negligable so long as we
don't start blocking in the main loop.